Add a screenController function that lets the UI interact with the game
We populate the board div with button using javascript and add an event listener
to them which when clicked does two things runs the playRound method and
accordingly updates the text content of those buttons, and changes the
turn message in top of the board. We also change the order of the if
statement in the playRound method so that it stops running if a winner
is found, there was also some wrong indentation in the method which is now
corrected. We also add some basic css to the project.